+++ lib/src/core/services/game_datasource.dart
@@ -1,7 +1,7 @@
 import 'package:fpdart/fpdart.dart';
 import 'package:my_app/src/core/exceptions.dart';
-import 'package:my_app/src/features/game/domain/game.dart';
-import 'package:my_app/src/features/game/domain/game_status.dart';
+import 'package:my_app/src/features/game/data/model/game.dart';
+import 'package:my_app/src/features/game/data/model/game_status.dart';
 import 'package:my_app/src/features/player/data/model/player.dart';
 
 abstract class GameDataSource {
