+++ lib/src/features/game/data/model/attempt.dart
@@ -1,6 +1,6 @@
 import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:my_app/src/core/supabase/table_interface.dart';
-import 'package:my_app/src/features/game/domain/game.dart';
+import 'package:my_app/src/features/game/data/model/game.dart';
 import 'package:my_app/src/features/player/data/model/player.dart';
 
 part 'attempt.g.dart';
